type: move and fix imports
