+++ src/contexts/ThemeProvider.tsx
@@ -1,16 +1,10 @@
 import { createContext, useEffect, useState } from 'react'
-
-type Theme = 'dark' | 'light' | 'system'
+import type { Theme, ThemeProviderState } from '@/types'
 
 interface ThemeProviderProps {
   children: React.ReactNode
   defaultTheme?: Theme
   storageKey?: string
-}
-
-export interface ThemeProviderState {
-  theme: Theme
-  setTheme: (theme: Theme) => void
 }
 
 const initialState: ThemeProviderState = {
