+++ src/contexts/index.tsx
@@ -1,2 +1,2 @@
 export { NavigationProvider, NavigationContext } from './NavigationProvider'
-export { ThemeProvider } from './ThemeProvider'
+export { ThemeProvider, ThemeProviderContext } from './ThemeProvider'
